requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
redis==5.0.5
pydantic==1.10.12
python-dotenv==1.0.0